This is my first stab at an algorithm for grouping the tests by their duration using the shard count provided by the user.

In this PR I am modifying the test method list to include duration time and then creating shard buckets based on the existing algorithm and comparing it to a time based approach, the algorithm for this was based on the Java implementation of Flank found here:
https://github.com/TestArmada/flank/blob/551818a0e0a8e054113aea5f98248ad436c9470f/Flank/src/main/java/com/walmart/otto/shards/ShardCreator.java

This version of the algorithm works by setting a bucket size for each shard equal to the longest test run in the list of tests. It then iterates through all tests in descending order attempting to fill each bucket with as many tests as possible. If all buckets are "full" and we have tests remaining, it starts again, with additional bucket space being equal to the remaining longest test.

Here is an example of the results:
WITH TIME BASED SORTING
Slowest shard = 191s
Total Time = 4575s // Time for all tests to be ran synchronously
WITHOUT TIME SORTING
Slowest shard = 329s
Total Time = 4575s // Time for all tests to be ran synchronously

Bear in mind that the algorithms performance is heavily based on the given shard count, number of tests, and their distribution of durations. This example was done with against 30 shards, with 70 tests and time range of 10-120 seconds.

I'd love to get some feedback on this approach, any C&C would be greatly appreciated!