@inktomi 2018-07-11
I think this is a fine first version of the algorithim. I worry a bit about the number of buckets that might be required - our project has around a thousand tests and given that the longest bucket will contain a single test (whatever test runs longest) we will probably end up with hundreds of buckets for the optimal case. 

Do you know if most projects have a handful / single of really long tests? That seems like an ideal case for this algorithm. Our suite plateaus pretty fast with no huge standout tests.

It would be really neat to be able to plug in algorithms. People could use the provided one, or configure their own based on their needs.