@bootstraponline 2018-07-11
> It would be really neat to be able to plug in algorithms. People could use the provided one, or configure their own based on their needs.

Yeah, I think that makes sense. There are unique situations where a custom solution is going to be the best fit. 

> I worry a bit about the number of buckets that might be required

For clarity, the algorithm in this PR is using a user specified bucket maximum. `testShards` is how you control the maximum amount of buckets being created.