@bootstraponline 2018-07-11
@inktomi does https://github.com/TestArmada/flank/issues/44 solve your use case?