@inktomi 2018-07-11
Yea, it would! We currently use a custom test runner that works off a manifest like that. It's much easier to have Flank generate it, but this would give teams flexibility if needed.