@bootstraponline 2018-07-11
Cool. I expect to resume work on Flank in August. I think reading in a list of tests may make sense once the other major changes have landed.