@bootstraponline 2018-09-06
This feature is on pause until the new architecture work is complete.

- https://github.com/TestArmada/flank/issues/286